- (snmpd.c, context_parse.c, party_parse.c, system.[ch]):   - 64bit fix: get_myaddr() should return int not long.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@781 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/snmplib/party_parse.c
@@ -90,7 +90,7 @@ read_party_database(filename)
     int privPublicLength, authPublicLength;
     char name[64];	/* friendly name */
     struct partyEntry *pp, *rp;
-    u_long myaddr;
+    u_int myaddr;
     int domain;
 
     fp = fopen(filename, "r");
